feat(svgo): add WASM binding + fix test signatures in turndown/svgo cores
svgo: extracts the quick-xml-based SVG optimizer (8 preset-default
plugins, multipass loop, byte-level result reporting) into
`_svgo-core`. The napi crate is now a thin shim; WASM exposes
`optimize(svg, config)` / `optimizeMany(svgs, config)` returning a
serde-shaped result object.

Also fixes test-rust failure: the test modules in `_svgo-core/src/lib.rs`
and `_turndown-core/src/lib.rs` still called the old napi-shaped
`optimize(String, Option<Config>)` / `turndown(String, Option<Opts>)`
signatures after the core split; updated them to `&str + &Config`
to match the new core APIs.

Adds `serde_json = "1"` dev-dep to svgo/wasm/Cargo.toml — required
by the wasm-bindgen-test fixtures.
