+++ docs/readmes/svgo.html
@@ -65,5 +65,10 @@ static assets. Path-arithmetic plugins (<code>convertPathData</code>,
 FFI-crossing and destroy the performance thesis.</p>
 <p>See <a href="./__conformance__/divergences.md"><code>__conformance__/divergences.md</code></a>
 for known byte-level differences vs. upstream.</p>
+<h2 id="install-for-the-browser">Install for the browser</h2>
+<p>The same <code>import</code> works in Angular, React, Vite, esbuild, and webpack ≥ 5 — the bundler picks the WASM build via the <code>browser</code> conditional export:</p>
+<pre><code class="language-ts">import { optimize } from '@amigo-labs/svgo'
+</code></pre>
+<p><code>quick-xml</code> is small (~80 KB gzipped) — comfortably under the 500 KB browser budget. Useful for build-time SVG optimization in client-side workflows (e.g., dynamic icon-sprite generation).</p>
 <h2 id="license">License</h2>
 <p>MIT</p>
